// Map311/generated/ManifestTypes.d.ts
/*
*This is auto generated from the ControlManifest.Input.xml file
*/

// Define IInputs and IOutputs Type. They should match with ControlManifest.
export interface IInputs {
    defaultLatitude: ComponentFramework.PropertyTypes.DecimalNumberProperty;
    defaultLongitude: ComponentFramework.PropertyTypes.DecimalNumberProperty;
    defaultZoom: ComponentFramework.PropertyTypes.WholeNumberProperty;
    sampleDataSet: ComponentFramework.PropertyTypes.DataSet;
}
export interface IOutputs {
}